+++ common/src/main/java/com/majruszsaccessories/accessories/AdventurerKit.java
@@ -5,10 +5,10 @@ import com.majruszlibrary.data.Reader;
 import com.majruszlibrary.math.Range;
 import com.majruszsaccessories.MajruszsAccessories;
 import com.majruszsaccessories.accessories.components.MoreChestLoot;
-import com.majruszsaccessories.accessories.components.TradeOffer;
 import com.majruszsaccessories.common.AccessoryHandler;
 import com.majruszsaccessories.common.BonusComponent;
 import com.majruszsaccessories.common.BonusHandler;
+import com.majruszsaccessories.common.components.TradeOffer;
 import com.majruszsaccessories.events.base.CustomConditions;
 import com.majruszsaccessories.items.AccessoryItem;
 
@@ -19,7 +19,7 @@ public class AdventurerKit extends AccessoryHandler {
 
 		this.add( MoreChestLoot.create( 1.2f ) )
 			.add( AnyChestDropChance.create() )
-			.add( TradeOffer.create( 7 ) );
+			.add( TradeOffer.create( MajruszsAccessories.GAMBLING_CARD, 1 ) );
 	}
 
 	static class AnyChestDropChance extends BonusComponent< AccessoryItem > {