+++ common/src/main/java/com/majruszsaccessories/accessories/AdventurerRune.java
@@ -5,7 +5,6 @@ import com.majruszsaccessories.MajruszsAccessories;
 import com.majruszsaccessories.accessories.components.BrushingExtraItem;
 import com.majruszsaccessories.accessories.components.MoreChestLoot;
 import com.majruszsaccessories.accessories.components.SwimmingSpeedBonus;
-import com.majruszsaccessories.accessories.components.TradeOffer;
 import com.majruszsaccessories.common.AccessoryHandler;
 
 @AutoInstance
@@ -15,7 +14,6 @@ public class AdventurerRune extends AccessoryHandler {
 
 		this.add( MoreChestLoot.create( 1.5f ) )
 			.add( BrushingExtraItem.create( 0.2f ) )
-			.add( SwimmingSpeedBonus.create( 0.25f ) )
-			.add( TradeOffer.create( 17 ) );
+			.add( SwimmingSpeedBonus.create( 0.25f ) );
 	}
 }