+++ common/src/main/java/com/majruszsaccessories/accessories/HouseholdRune.java
@@ -4,7 +4,6 @@ import com.majruszlibrary.annotation.AutoInstance;
 import com.majruszsaccessories.MajruszsAccessories;
 import com.majruszsaccessories.accessories.components.SleepingBonuses;
 import com.majruszsaccessories.accessories.components.StrongerPotions;
-import com.majruszsaccessories.accessories.components.TradeOffer;
 import com.majruszsaccessories.accessories.components.TradingDiscount;
 import com.majruszsaccessories.common.AccessoryHandler;
 
@@ -15,7 +14,6 @@ public class HouseholdRune extends AccessoryHandler {
 
 		this.add( TradingDiscount.create( 0.15f ) )
 			.add( SleepingBonuses.create( 1, 360 ) )
-			.add( StrongerPotions.create( 0.5f, 1 ) )
-			.add( TradeOffer.create( 17 ) );
+			.add( StrongerPotions.create( 0.5f, 1 ) );
 	}
 }