+++ common/src/main/java/com/majruszsaccessories/accessories/MinerGuide.java
@@ -6,10 +6,10 @@ import com.majruszlibrary.events.OnLootGenerated;
 import com.majruszlibrary.math.Range;
 import com.majruszsaccessories.MajruszsAccessories;
 import com.majruszsaccessories.accessories.components.MiningSpeedBonus;
-import com.majruszsaccessories.accessories.components.TradeOffer;
 import com.majruszsaccessories.common.AccessoryHandler;
 import com.majruszsaccessories.common.BonusComponent;
 import com.majruszsaccessories.common.BonusHandler;
+import com.majruszsaccessories.common.components.TradeOffer;
 import com.majruszsaccessories.events.base.CustomConditions;
 import com.majruszsaccessories.items.AccessoryItem;
 
@@ -20,7 +20,7 @@ public class MinerGuide extends AccessoryHandler {
 
 		this.add( MiningSpeedBonus.create( 0.1f ) )
 			.add( UndergroundChestDropChance.create() )
-			.add( TradeOffer.create( 7 ) );
+			.add( TradeOffer.create( MajruszsAccessories.GAMBLING_CARD, 1 ) );
 	}
 
 	static class UndergroundChestDropChance extends BonusComponent< AccessoryItem > {