+++ common/src/main/java/com/majruszsaccessories/accessories/MinerRune.java
@@ -5,7 +5,6 @@ import com.majruszsaccessories.MajruszsAccessories;
 import com.majruszsaccessories.accessories.components.MiningDurabilityBonus;
 import com.majruszsaccessories.accessories.components.MiningExtraItem;
 import com.majruszsaccessories.accessories.components.MiningSpeedBonus;
-import com.majruszsaccessories.accessories.components.TradeOffer;
 import com.majruszsaccessories.common.AccessoryHandler;
 
 @AutoInstance
@@ -15,7 +14,6 @@ public class MinerRune extends AccessoryHandler {
 
 		this.add( MiningExtraItem.create( 0.04f ) )
 			.add( MiningSpeedBonus.create( 0.12f ) )
-			.add( MiningDurabilityBonus.create( 0.12f ) )
-			.add( TradeOffer.create( 17 ) );
+			.add( MiningDurabilityBonus.create( 0.12f ) );
 	}
 }