+++ common/src/main/java/com/majruszsaccessories/accessories/TamedPotatoBeetle.java
@@ -6,10 +6,10 @@ import com.majruszlibrary.events.base.Events;
 import com.majruszlibrary.math.Range;
 import com.majruszsaccessories.MajruszsAccessories;
 import com.majruszsaccessories.accessories.components.HarvestingDoubleCrops;
-import com.majruszsaccessories.accessories.components.TradeOffer;
 import com.majruszsaccessories.common.AccessoryHandler;
 import com.majruszsaccessories.common.BonusComponent;
 import com.majruszsaccessories.common.BonusHandler;
+import com.majruszsaccessories.common.components.TradeOffer;
 import com.majruszsaccessories.events.OnAccessoryDropChanceGet;
 import com.majruszsaccessories.items.AccessoryItem;
 import net.minecraft.world.level.block.Blocks;
@@ -21,7 +21,7 @@ public class TamedPotatoBeetle extends AccessoryHandler {
 
 		this.add( HarvestingDoubleCrops.create( 0.25f ) )
 			.add( HarvestingDropChance.create() )
-			.add( TradeOffer.create( 7 ) );
+			.add( TradeOffer.create( MajruszsAccessories.GAMBLING_CARD, 1 ) );
 	}
 
 	static class HarvestingDropChance extends BonusComponent< AccessoryItem > {