+++ common/src/main/java/com/majruszsaccessories/accessories/UnbreakableFishingLine.java
@@ -3,8 +3,8 @@ package com.majruszsaccessories.accessories;
 import com.majruszlibrary.annotation.AutoInstance;
 import com.majruszsaccessories.MajruszsAccessories;
 import com.majruszsaccessories.accessories.components.FishingLureBonus;
-import com.majruszsaccessories.accessories.components.TradeOffer;
 import com.majruszsaccessories.common.AccessoryHandler;
+import com.majruszsaccessories.common.components.TradeOffer;
 
 @AutoInstance
 public class UnbreakableFishingLine extends AccessoryHandler {
@@ -14,6 +14,6 @@ public class UnbreakableFishingLine extends AccessoryHandler {
 		this.add( FishingLureBonus.create( 0.2f ) )
 			.add( AnglerTrophy.FishingDropChance.create( 0.005f ) )
 			.add( AnglerTrophy.FishDropChance.create( 0.005f ) )
-			.add( TradeOffer.create( 7 ) );
+			.add( TradeOffer.create( MajruszsAccessories.GAMBLING_CARD, 1 ) );
 	}
 }