+++ changelog.md
@@ -1,5 +1,6 @@
 - added Gambling Card
 - added Reverse Card
 - removed Jadeite
+- reworked Wandering Trader trades for accessories and boosters
 - optimized accessory checks
 - fixed compatibility crash with MineColonies (reported by @Ryus Lightworks)